pyproject.toml
[build-system]
requires = ["setuptools>=61", "setuptools_scm[toml]>=6.2"] 
build-backend = "setuptools.build_meta"

[project]
name = "McPainMKV"
dynamic = ["version"]
description = "My personal mkv converter for my media server."
authors = [
    { name="TheGreatMcPain", email="james@thegreatmcpain.xyz" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg-normalize>=1.37.3",
    "iso639-lang>=2.6.3",
    "makemkv>=0.3.1",
    "psutil>=7.2.2",
    "rich>=14.3.2",
    "subtitle-filter>=1.5.0",
    "vapoursynth>=73",
    "vsstubs>=1.0.2",
]

[project.scripts]
mcpain-mkv = "mcpainmkv:main"

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "dirty-tag"
